+++ flutter/lib/consts.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/material.dart';
+
 const double kDesktopRemoteTabBarHeight = 28.0;
 
 /// [kAppTypeMain] used by 'Desktop Main Page' , 'Mobile (Client and Server)' , 'Desktop CM Page'
@@ -16,6 +18,8 @@ const int kDesktopDefaultDisplayWidth = 1080;
 const int kDesktopDefaultDisplayHeight = 720;
 
 const kInvalidValueStr = "InvalidValueStr";
+
+const kMobilePageConstraints = BoxConstraints(maxWidth: 600);
 
 /// flutter/packages/flutter/lib/src/services/keyboard_key.dart -> _keyLabels
 /// see [LogicalKeyboardKey.keyLabel]
