+++ flutter/lib/desktop/pages/connection_page.dart
@@ -1,17 +1,15 @@
 import 'dart:async';
 import 'dart:convert';
 
-import 'package:contextmenu/contextmenu.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_hbb/desktop/pages/desktop_home_page.dart';
-import 'package:flutter_hbb/desktop/widgets/peer_widget.dart';
-import 'package:flutter_hbb/desktop/widgets/peercard_widget.dart';
-import 'package:flutter_hbb/utils/multi_window_manager.dart';
+import 'package:flutter_hbb/common/widgets/address_book.dart';
 import 'package:get/get.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
 import '../../common.dart';
 import '../../common/formatter/id_formatter.dart';
+import '../../common/widgets/peer_tab_page.dart';
+import '../../common/widgets/peer_widget.dart';
 import '../../models/platform_model.dart';
 
 /// Connection page for connecting to a remote peer.
@@ -66,7 +64,7 @@ class _ConnectionPageState extends State<ConnectionPage> {
                   SizedBox(height: 12),
                   Divider(),
                   Expanded(
-                      child: _PeerTabbedPage(
+                      child: PeerTabPage(
                     tabs: [
                       translate('Recent Sessions'),
                       translate('Favorites'),
@@ -77,15 +75,7 @@ class _ConnectionPageState extends State<ConnectionPage> {
                       RecentPeerWidget(),
                       FavoritePeerWidget(),
                       DiscoveredPeerWidget(),
-                      FutureBuilder<Widget>(
-                          future: buildAddressBook(context),
-                          builder: (context, snapshot) {
-                            if (snapshot.hasData) {
-                              return snapshot.data!;
-                            } else {
-                              return const Offstage();
-                            }
-                          }),
+                      const AddressBook(),
                     ],
                   )),
                 ],
@@ -102,23 +92,7 @@ class _ConnectionPageState extends State<ConnectionPage> {
   /// Connects to the selected peer.
   void onConnect({bool isFileTransfer = false}) {
     final id = _idController.id;
-    connect(id, isFileTransfer: isFileTransfer);
-  }
-
-  /// Connect to a peer with [id].
-  /// If [isFileTransfer], starts a session only for file transfer.
-  void connect(String id, {bool isFileTransfer = false}) async {
-    if (id == '') return;
-    id = id.replaceAll(' ', '');
-    if (isFileTransfer) {
-      await rustDeskWinManager.newFileTransfer(id);
-    } else {
-      await rustDeskWinManager.newRemoteDesktop(id);
-    }
-    FocusScopeNode currentFocus = FocusScope.of(context);
-    if (!currentFocus.hasPrimaryFocus) {
-      currentFocus.unfocus();
-    }
+    connect(context, id, isFileTransfer: isFileTransfer);
   }
 
   /// UI for the search bar.
@@ -371,605 +345,5 @@ class _ConnectionPageState extends State<ConnectionPage> {
         jsonDecode(await bind.mainGetConnectStatus()) as Map<String, dynamic>;
     svcStatusCode.value = status["status_num"];
     svcIsUsingPublicServer.value = await bind.mainIsUsingPublicServer();
-  }
-
-  handleLogin() {
-    loginDialog().then((success) {
-      if (success) {
-        setState(() {});
-      }
-    });
-  }
-
-  Future<Widget> buildAddressBook(BuildContext context) async {
-    final token = await bind.mainGetLocalOption(key: 'access_token');
-    if (token.trim().isEmpty) {
-      return Center(
-        child: InkWell(
-          onTap: handleLogin,
-          child: Text(
-            translate("Login"),
-            style: TextStyle(decoration: TextDecoration.underline),
-          ),
-        ),
-      );
-    }
-    final model = gFFI.abModel;
-    return FutureBuilder(
-        future: model.getAb(),
-        builder: (context, snapshot) {
-          if (snapshot.hasData) {
-            return _buildAddressBook(context);
-          } else if (snapshot.hasError) {
-            return Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Text(translate("${snapshot.error}")),
-                TextButton(
-                    onPressed: () {
-                      setState(() {});
-                    },
-                    child: Text(translate("Retry")))
-              ],
-            );
-          } else {
-            if (model.abLoading) {
-              return const Center(
-                child: CircularProgressIndicator(),
-              );
-            } else if (model.abError.isNotEmpty) {
-              return Center(
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Text(translate("${model.abError}")),
-                    TextButton(
-                        onPressed: () {
-                          setState(() {});
-                        },
-                        child: Text(translate("Retry")))
-                  ],
-                ),
-              );
-            } else {
-              return Offstage();
-            }
-          }
-        });
-  }
-
-  Widget _buildAddressBook(BuildContext context) {
-    return Row(
-      children: [
-        Card(
-          shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(20),
-              side: BorderSide(color: MyTheme.grayBg)),
-          child: Container(
-            width: 200,
-            height: double.infinity,
-            padding: EdgeInsets.symmetric(horizontal: 12.0, vertical: 8.0),
-            child: Column(
-              children: [
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Text(translate('Tags')),
-                    InkWell(
-                      child: PopupMenuButton(
-                          itemBuilder: (context) => [
-                                PopupMenuItem(
-                                  child: Text(translate("Add ID")),
-                                  value: 'add-id',
-                                ),
-                                PopupMenuItem(
-                                  child: Text(translate("Add Tag")),
-                                  value: 'add-tag',
-                                ),
-                                PopupMenuItem(
-                                  child: Text(translate("Unselect all tags")),
-                                  value: 'unset-all-tag',
-                                ),
-                              ],
-                          onSelected: handleAbOp,
-                          child: Icon(Icons.more_vert_outlined)),
-                    )
-                  ],
-                ),
-                Expanded(
-                  child: Container(
-                    width: double.infinity,
-                    height: double.infinity,
-                    decoration: BoxDecoration(
-                        border: Border.all(color: MyTheme.darkGray)),
-                    child: Obx(
-                      () => Wrap(
-                        children: gFFI.abModel.tags
-                            .map((e) => buildTag(e, gFFI.abModel.selectedTags,
-                                    onTap: () {
-                                  //
-                                  if (gFFI.abModel.selectedTags.contains(e)) {
-                                    gFFI.abModel.selectedTags.remove(e);
-                                  } else {
-                                    gFFI.abModel.selectedTags.add(e);
-                                  }
-                                }))
-                            .toList(),
-                      ),
-                    ),
-                  ).marginSymmetric(vertical: 8.0),
-                )
-              ],
-            ),
-          ),
-        ).marginOnly(right: 8.0),
-        Expanded(
-          child: Align(
-              alignment: Alignment.topLeft, child: AddressBookPeerWidget()),
-        )
-      ],
-    );
-  }
-
-  Widget buildTag(String tagName, RxList<dynamic> rxTags, {Function()? onTap}) {
-    return ContextMenuArea(
-      width: 100,
-      builder: (context) => [
-        ListTile(
-          title: Text(translate("Delete")),
-          onTap: () {
-            gFFI.abModel.deleteTag(tagName);
-            gFFI.abModel.updateAb();
-            Future.delayed(Duration.zero, () => Get.back());
-          },
-        )
-      ],
-      child: GestureDetector(
-        onTap: onTap,
-        child: Obx(
-          () => Container(
-            decoration: BoxDecoration(
-                color: rxTags.contains(tagName) ? Colors.blue : null,
-                border: Border.all(color: MyTheme.darkGray),
-                borderRadius: BorderRadius.circular(10)),
-            margin: EdgeInsets.symmetric(horizontal: 4.0, vertical: 8.0),
-            padding: EdgeInsets.symmetric(vertical: 2.0, horizontal: 8.0),
-            child: Text(
-              tagName,
-              style: TextStyle(
-                  color: rxTags.contains(tagName) ? MyTheme.white : null),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  /// tag operation
-  void handleAbOp(String value) {
-    if (value == 'add-id') {
-      abAddId();
-    } else if (value == 'add-tag') {
-      abAddTag();
-    } else if (value == 'unset-all-tag') {
-      gFFI.abModel.unsetSelectedTags();
-    }
-  }
-
-  void abAddId() async {
-    var field = "";
-    var msg = "";
-    var isInProgress = false;
-    TextEditingController controller = TextEditingController(text: field);
-
-    gFFI.dialogManager.show((setState, close) {
-      submit() async {
-        setState(() {
-          msg = "";
-          isInProgress = true;
-        });
-        field = controller.text.trim();
-        if (field.isEmpty) {
-          // pass
-        } else {
-          final ids = field.trim().split(RegExp(r"[\s,;\n]+"));
-          field = ids.join(',');
-          for (final newId in ids) {
-            if (gFFI.abModel.idContainBy(newId)) {
-              continue;
-            }
-            gFFI.abModel.addId(newId);
-          }
-          await gFFI.abModel.updateAb();
-          this.setState(() {});
-          // final currentPeers
-        }
-        close();
-      }
-
-      return CustomAlertDialog(
-        title: Text(translate("Add ID")),
-        content: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(translate("whitelist_sep")),
-            const SizedBox(
-              height: 8.0,
-            ),
-            Row(
-              children: [
-                Expanded(
-                  child: TextField(
-                      maxLines: null,
-                      decoration: InputDecoration(
-                        border: const OutlineInputBorder(),
-                        errorText: msg.isEmpty ? null : translate(msg),
-                      ),
-                      controller: controller,
-                      focusNode: FocusNode()..requestFocus()),
-                ),
-              ],
-            ),
-            const SizedBox(
-              height: 4.0,
-            ),
-            Offstage(
-                offstage: !isInProgress, child: const LinearProgressIndicator())
-          ],
-        ),
-        actions: [
-          TextButton(onPressed: close, child: Text(translate("Cancel"))),
-          TextButton(onPressed: submit, child: Text(translate("OK"))),
-        ],
-        onSubmit: submit,
-        onCancel: close,
-      );
-    });
-  }
-
-  void abAddTag() async {
-    var field = "";
-    var msg = "";
-    var isInProgress = false;
-    TextEditingController controller = TextEditingController(text: field);
-    gFFI.dialogManager.show((setState, close) {
-      submit() async {
-        setState(() {
-          msg = "";
-          isInProgress = true;
-        });
-        field = controller.text.trim();
-        if (field.isEmpty) {
-          // pass
-        } else {
-          final tags = field.trim().split(RegExp(r"[\s,;\n]+"));
-          field = tags.join(',');
-          for (final tag in tags) {
-            gFFI.abModel.addTag(tag);
-          }
-          await gFFI.abModel.updateAb();
-          // final currentPeers
-        }
-        close();
-      }
-
-      return CustomAlertDialog(
-        title: Text(translate("Add Tag")),
-        content: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(translate("whitelist_sep")),
-            const SizedBox(
-              height: 8.0,
-            ),
-            Row(
-              children: [
-                Expanded(
-                  child: TextField(
-                    maxLines: null,
-                    decoration: InputDecoration(
-                      border: const OutlineInputBorder(),
-                      errorText: msg.isEmpty ? null : translate(msg),
-                    ),
-                    controller: controller,
-                    focusNode: FocusNode()..requestFocus(),
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(
-              height: 4.0,
-            ),
-            Offstage(
-                offstage: !isInProgress, child: const LinearProgressIndicator())
-          ],
-        ),
-        actions: [
-          TextButton(onPressed: close, child: Text(translate("Cancel"))),
-          TextButton(onPressed: submit, child: Text(translate("OK"))),
-        ],
-        onSubmit: submit,
-        onCancel: close,
-      );
-    });
-  }
-
-  void abEditTag(String id) {
-    var isInProgress = false;
-
-    final tags = List.of(gFFI.abModel.tags);
-    var selectedTag = gFFI.abModel.getPeerTags(id).obs;
-
-    gFFI.dialogManager.show((setState, close) {
-      submit() async {
-        setState(() {
-          isInProgress = true;
-        });
-        gFFI.abModel.changeTagForPeer(id, selectedTag);
-        await gFFI.abModel.updateAb();
-        close();
-      }
-
-      return CustomAlertDialog(
-        title: Text(translate("Edit Tag")),
-        content: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Container(
-              padding:
-                  const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
-              child: Wrap(
-                children: tags
-                    .map((e) => buildTag(e, selectedTag, onTap: () {
-                          if (selectedTag.contains(e)) {
-                            selectedTag.remove(e);
-                          } else {
-                            selectedTag.add(e);
-                          }
-                        }))
-                    .toList(growable: false),
-              ),
-            ),
-            Offstage(
-                offstage: !isInProgress, child: const LinearProgressIndicator())
-          ],
-        ),
-        actions: [
-          TextButton(onPressed: close, child: Text(translate("Cancel"))),
-          TextButton(onPressed: submit, child: Text(translate("OK"))),
-        ],
-        onSubmit: submit,
-        onCancel: close,
-      );
-    });
-  }
-}
-
-class _PeerTabbedPage extends StatefulWidget {
-  final List<String> tabs;
-  final List<Widget> children;
-  const _PeerTabbedPage({required this.tabs, required this.children, Key? key})
-      : super(key: key);
-  @override
-  _PeerTabbedPageState createState() => _PeerTabbedPageState();
-}
-
-class _PeerTabbedPageState extends State<_PeerTabbedPage>
-    with SingleTickerProviderStateMixin {
-  late PageController _pageController = PageController();
-  RxInt _tabIndex = 0.obs;
-
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  // hard code for now
-  void _handleTabSelection(int index) {
-    // reset search text
-    peerSearchText.value = "";
-    peerSearchTextController.clear();
-    _tabIndex.value = index;
-    _pageController.jumpToPage(index);
-    switch (index) {
-      case 0:
-        bind.mainLoadRecentPeers();
-        break;
-      case 1:
-        bind.mainLoadFavPeers();
-        break;
-      case 2:
-        bind.mainDiscover();
-        break;
-      case 3:
-        gFFI.abModel.updateAb();
-        break;
-    }
-  }
-
-  @override
-  void dispose() {
-    _pageController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      textBaseline: TextBaseline.ideographic,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Container(
-          height: 28,
-          child: Row(
-            children: [
-              Expanded(child: _createTabBar(context)),
-              _createSearchBar(context),
-              _createPeerViewTypeSwitch(context),
-            ],
-          ),
-        ),
-        _createTabBarView(),
-      ],
-    );
-  }
-
-  Widget _createTabBar(BuildContext context) {
-    return ListView(
-        scrollDirection: Axis.horizontal,
-        shrinkWrap: true,
-        controller: ScrollController(),
-        children: super.widget.tabs.asMap().entries.map((t) {
-          return Obx(() => GestureDetector(
-                child: Container(
-                    padding: EdgeInsets.symmetric(horizontal: 8),
-                    decoration: BoxDecoration(
-                      color: _tabIndex.value == t.key
-                          ? MyTheme.color(context).bg
-                          : null,
-                      borderRadius: BorderRadius.circular(2),
-                    ),
-                    child: Align(
-                      alignment: Alignment.center,
-                      child: Text(
-                        t.value,
-                        textAlign: TextAlign.center,
-                        style: TextStyle(
-                            height: 1,
-                            fontSize: 14,
-                            color: _tabIndex.value == t.key
-                                ? MyTheme.color(context).text
-                                : MyTheme.color(context).lightText),
-                      ),
-                    )),
-                onTap: () => _handleTabSelection(t.key),
-              ));
-        }).toList());
-  }
-
-  Widget _createTabBarView() {
-    return Expanded(
-        child: PageView(
-                controller: _pageController, children: super.widget.children)
-            .marginSymmetric(vertical: 12));
-  }
-
-  _createSearchBar(BuildContext context) {
-    RxBool focused = false.obs;
-    FocusNode focusNode = FocusNode();
-    focusNode.addListener(() => focused.value = focusNode.hasFocus);
-    RxBool rowHover = false.obs;
-    RxBool clearHover = false.obs;
-    return Container(
-      width: 120,
-      height: 25,
-      margin: EdgeInsets.only(right: 13),
-      decoration: BoxDecoration(color: MyTheme.color(context).bg),
-      child: Obx(() => Row(
-            children: [
-              Expanded(
-                child: MouseRegion(
-                  onEnter: (_) => rowHover.value = true,
-                  onExit: (_) => rowHover.value = false,
-                  child: Row(
-                    children: [
-                      Icon(
-                        IconFont.search,
-                        size: 16,
-                        color: MyTheme.color(context).placeholder,
-                      ).marginSymmetric(horizontal: 4),
-                      Expanded(
-                        child: TextField(
-                          controller: peerSearchTextController,
-                          onChanged: (searchText) {
-                            peerSearchText.value = searchText;
-                          },
-                          focusNode: focusNode,
-                          textAlign: TextAlign.start,
-                          maxLines: 1,
-                          cursorColor: MyTheme.color(context).lightText,
-                          cursorHeight: 18,
-                          cursorWidth: 1,
-                          style: TextStyle(fontSize: 14),
-                          decoration: InputDecoration(
-                            contentPadding: EdgeInsets.symmetric(vertical: 6),
-                            hintText:
-                                focused.value ? null : translate("Search ID"),
-                            hintStyle: TextStyle(
-                                fontSize: 14,
-                                color: MyTheme.color(context).placeholder),
-                            border: InputBorder.none,
-                            isDense: true,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              Offstage(
-                offstage: !(peerSearchText.value.isNotEmpty &&
-                    (rowHover.value || clearHover.value)),
-                child: InkWell(
-                    onHover: (value) => clearHover.value = value,
-                    child: Icon(
-                      IconFont.round_close,
-                      size: 16,
-                      color: clearHover.value
-                          ? MyTheme.color(context).text
-                          : MyTheme.color(context).placeholder,
-                    ).marginSymmetric(horizontal: 4),
-                    onTap: () {
-                      peerSearchTextController.clear();
-                      peerSearchText.value = "";
-                    }),
-              )
-            ],
-          )),
-    );
-  }
-
-  _createPeerViewTypeSwitch(BuildContext context) {
-    final activeDeco = BoxDecoration(color: MyTheme.color(context).bg);
-    return Row(
-      children: [
-        Obx(
-          () => Container(
-            padding: EdgeInsets.all(4.0),
-            decoration:
-                peerCardUiType.value == PeerUiType.grid ? activeDeco : null,
-            child: InkWell(
-                onTap: () {
-                  peerCardUiType.value = PeerUiType.grid;
-                },
-                child: Icon(
-                  Icons.grid_view_rounded,
-                  size: 18,
-                  color: peerCardUiType.value == PeerUiType.grid
-                      ? MyTheme.color(context).text
-                      : MyTheme.color(context).lightText,
-                )),
-          ),
-        ),
-        Obx(
-          () => Container(
-            padding: EdgeInsets.all(4.0),
-            decoration:
-                peerCardUiType.value == PeerUiType.list ? activeDeco : null,
-            child: InkWell(
-                onTap: () {
-                  peerCardUiType.value = PeerUiType.list;
-                },
-                child: Icon(
-                  Icons.list,
-                  size: 18,
-                  color: peerCardUiType.value == PeerUiType.list
-                      ? MyTheme.color(context).text
-                      : MyTheme.color(context).lightText,
-                )),
-          ),
-        ),
-      ],
-    );
   }
 }
